requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
anthropic==0.77.0
python-dotenv==1.0.1
langchain
langchain-community
chromadb
sentence-transformers
beautifulsoup4
requests
playwright
